Validaciones y cambio de tipo de atributos

+++ DataAccess/Entities/TB_PRODUCTO.cs
@@ -21,13 +21,13 @@ namespace DataAccess.Entities
         public string Titulo { get; set; }
         public int Cantidad { get; set; }
         public string Caracteristicas { get; set; }
-        public decimal Precio { get; set; }
-        public decimal Coste { get; set; }
-        public decimal Margen { get; set; }
-        public decimal Descuento { get; set; }
-        public decimal PVP { get; set; }
+        public float Precio { get; set; }
+        public float Coste { get; set; }
+        public float Margen { get; set; }
+        public float Descuento { get; set; }
+        public float PVP { get; set; }
         public float IVA { get; set; }
-        public decimal Total { get; set; }
+        public float Total { get; set; }
 
     }
 }
